+++ slurmweb/apps/gateway.py
@@ -107,6 +107,9 @@ class SlurmwebAppGateway(SlurmwebWebApp, RFLTokenizedWebApp):
         SlurmwebAppRoute("/api/agents/<cluster>/stats", views.stats),
         SlurmwebAppRoute("/api/agents/<cluster>/metrics/<metric>", views.metrics),
         SlurmwebAppRoute("/api/agents/<cluster>/cache/stats", views.cache_stats),
+        SlurmwebAppRoute(
+            "/api/agents/<cluster>/cache/reset", views.cache_reset, methods=["POST"]
+        ),
         SlurmwebAppRoute("/api/agents/<cluster>/jobs", views.jobs),
         SlurmwebAppRoute("/api/agents/<cluster>/job/<int:job>", views.job),
         SlurmwebAppRoute("/api/agents/<cluster>/nodes", views.nodes),