+++ slurmweb/views/gateway.py
@@ -274,6 +274,12 @@ def cache_stats(cluster: str):
 
 @check_jwt
 @validate_cluster
+def cache_reset(cluster: str):
+    return proxy_agent(cluster, "cache/reset", request.token)
+
+
+@check_jwt
+@validate_cluster
 def jobs(cluster: str):
     return proxy_agent(cluster, "jobs", request.token)